finish add loadingbar

+++ src/app/core/module/students/students.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { StudentService } from '../../service/student.service';
 import { ToastrService } from 'ngx-toastr';
+import { LoadingBarService } from '@ngx-loading-bar/core';
 
 @Component({
   selector: 'app-students',
@@ -12,6 +13,7 @@ export class StudentsComponent implements OnInit {
   allStudents: [];
 
   constructor(
+    private loadingBarSvc: LoadingBarService,
     private router: Router,
     private studentSvc: StudentService,
     private toastrSvc: ToastrService
@@ -22,7 +24,9 @@ export class StudentsComponent implements OnInit {
   }
 
   private getAllStudentsData() {
+    this.loadingBarSvc.start();
     this.studentSvc.getAllStudents().subscribe(response => {
+      this.loadingBarSvc.complete();
       this.allStudents = response.data;
     });
   }
@@ -36,10 +40,15 @@ export class StudentsComponent implements OnInit {
   }
 
   public deleteStudents(studentId: string) {
-    this.studentSvc.deleteStudent(studentId).subscribe(response => {
-      this.toastrSvc.success('message', response.message);
-      this.getAllStudentsData();
+    if(confirm("Are you delete this?")) {
+      this.loadingBarSvc.start();
+      this.studentSvc.deleteStudent(studentId).subscribe(response => {
+        this.loadingBarSvc.complete();
+        this.toastrSvc.success('message', response.message);
+        this.getAllStudentsData();
     });
+    }
+    
   }
 
 }
